# src/Models/Product.cs
namespace FeaturesEnabled.ClaimBased.AutoSwagger.Api.Core.Models
{
    public class Product
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public string Description { get; set; }
    }
}